\name{CalcMI_time2}
\Rdversion{1.1}
\alias{CalcMI_time2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
CalcMI_time2
}
\description{Compute the d-delayed Mutual information}
\usage{
CalcMI_time2(l, t, delta)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{l}{
one gene profile
}
  \item{t}{
another gene profile
}
  \item{delta}{
maximum delay
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.