\name{DPI2_TDAracne}
\Rdversion{1.1}
\alias{DPI2_TDAracne}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
DPI2_TDAracne
}
\description{MAke the second DPI }
\usage{
DPI2_TDAracne(MItab, tolerance)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MItab}{
MItab is the adjacency matrix before the second DPI
}
  \item{tolerance}{
tolerance is the DPI tolerance.
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.