\name{DPI_TDAracne}
\Rdversion{1.1}
\alias{DPI_TDAracne}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
DPI_TDAracne
}
\description{MAake the first DPI}
\usage{
DPI_TDAracne(MItab, delta, tolerance)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MItab}{
MItab is the adjacency matrix before DPI
}
  \item{delta}{
delta is the maximum time delay allowed to infer connections.   
}
  \item{tolerance}{
tolerance is the DPI tolerance.
}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
