\name{IcEfx}
\Rdversion{1.1}
\alias{IcEfx}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
IcEfx
}
\description{Select the point of Initial change Expression of the genes}
\usage{
IcEfx(z, likehood, logarit)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{z}{
z is the data matrix
}
  \item{likehood}{
likehood is the fold change used as threshold to state the initial change expression
}
  \item{logarit}{
if z is log put logarithm == 0;
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.