\name{PercentileC}
\Rdversion{1.1}
\alias{PercentileC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
PercentileC
}
\description{Percentile row normalization, each column goes from 0 to 1}
\usage{
PercentileC(z, N)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{z}{
z is the data matrix
	}
  \item{N}{
N is respectively the number of bins in percentile normalization or in rank normalization
	}
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.