\name{plotRgraphviz}
\Rdversion{1.1}
\alias{plotRgraphviz}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plotRgraphviz
}
\description{use Rgraphviz to plot the adj; bonus help function}
\usage{
plotRgraphviz(Influence)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Influence}{
the adj matrix
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.