\name{saveTime}
\Rdversion{1.1}
\alias{saveTime}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
saveTime
}
\description{ make some useful check on the data}
\usage{
saveTime(newz, delta)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{newz}{
newz is the data matrix
}
  \item{delta}{
delta is the maximum time delay allowed to infer connections
}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
